requirements.txt
# Core
requests>=2.31.0
PyYAML>=6.0

# Web scraping
scrapling[all]>=0.4.2
readability-lxml>=0.8.1
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Async
aiohttp>=3.9.0
aiofiles>=23.0.0

# Utils
tqdm>=4.66.0
python-dateutil>=2.8.0
urllib3>=2.0.0
